biblioteka/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
matplotlib
pandas
python-dotenv
